Refactor SAAImpl into API implementation (kept in SAAImpl) and hardware device glue logic (now in SAADevice), and as part of this change: ; eliminate extra indirection and malloc/free/assert logic in connecting the submodules; eliminate unsigned short in favour of int; simplify/remove some conditional logic to just use math operators; simplify freq/noise/amp logic a little (freq now returns 0/1, just like noise does, instead of 0/2); and incorporate defns.h

+++ src/types.h
@@ -24,7 +24,7 @@ typedef struct
 {
 	int nNumberOfPhases;
 	bool bLooping;
-	unsigned short nLevels[2][2][16]; // [Resolution][Phase][Withinphase]
+	int nLevels[2][2][16]; // [Resolution][Phase][Withinphase]
 } ENVDATA;
 
 #ifdef WIN32
